+++ rascal-lsp/src/main/java/org/rascalmpl/vscode/lsp/dap/RascalDebugAdapter.java
@@ -28,6 +28,8 @@ import org.rascalmpl.values.RascalValueFactory;
 import org.rascalmpl.values.parsetrees.ITree;
 import org.rascalmpl.values.parsetrees.ProductionAdapter;
 import org.rascalmpl.values.parsetrees.TreeAdapter;
+import org.rascalmpl.vscode.lsp.dap.breakpoint.BreakpointsCollection;
+import org.rascalmpl.vscode.lsp.dap.variable.RascalVariable;
 import org.rascalmpl.vscode.lsp.terminal.LSPTerminalREPL;
 import org.rascalmpl.vscode.lsp.util.RascalServices;
